Fix date problem in L2 plots inherited from average calibration data frame.
Error somewhere in the code that takes spline selected points and averages
them that produces a row of all 0's in the output data frame. This error
hasn't been found - provisional fix here removes these rows before the data
frame is returned so that the monavgs plots don't include an additional point
at (0,0) for all six frames. This issue was raised in #30, but is only a band-aid
and doesn't fix the underlying problem...