Upgrades to peak finding function.
Modifications to peak finding function that seeks to separate standard
analysis using time and changes in H2O concentration. Significant improvement,
but still needs some work. Notably, some of the output peaks still have
large time gaps in them that need to be removed. This improves code creating issue #32, but it is not resolved.