Add explicit, clear diagnostic plot directory
Adds an explicitly specified plot directory tied
to the current time at code execution. Output is a little
wonky however, as colons are apparently not allowed in folder
output (though I'm confused as to why forward slashes are preferred,
since these are the folder dividers in os x to begin with!

Closes #36.