Update L2 files to match what was used in ACPD folder
At one point I broke off a "stable" version of the code to a different
folder, yet continued to make changes. Copying those changes in the L2
files over in this commit. The biggest changes were in the humidity cal,
which allowed for a hyberbolic offset parameterization. Adding these features
in before tweaking them yet again...